+++ TablePro/Core/Storage/Preferences/PreferenceKeys.swift
@@ -10,6 +10,7 @@ enum PreferenceKeys {
     static let linkedSQLFolders = DefaultsKey<[LinkedSQLFolder]>("com.TablePro.linkedSQLFolders")
     static let selectedSettingsPane = DefaultsKey<String>("com.TablePro.settings.selectedPane")
     static let rowInspectorJsonFieldHeight = DefaultsKey<Double>("com.TablePro.rightSidebar.jsonFieldHeight")
+    static let rowInspectorTextFieldHeight = DefaultsKey<Double>("com.TablePro.rightSidebar.textFieldHeight")
     static let workspaceRailOrder = DefaultsKey<[WorkspaceID]>("com.TablePro.workspaceRail.order")
     static let queryPlanRawFontSize = DefaultsKey<Double>("com.TablePro.queryPlan.rawFontSize")
 
@@ -18,6 +19,7 @@ enum PreferenceKeys {
         linkedSQLFolders.name,
         selectedSettingsPane.name,
         rowInspectorJsonFieldHeight.name,
+        rowInspectorTextFieldHeight.name,
         workspaceRailOrder.name,
         queryPlanRawFontSize.name,
     ]